Construct xrt::bo associated with hwctx (#7165)
Add xrt::bo constructors that take hwctx.  The goal is to remove the
need for encoding 'slotidx' in the grpid flags passed during
construction and instead make the context explicit during
construction.

This PR does not remove the slotidx encoding, this has to be a
coordinated change between XRT and various shim implementations so
this is for follow up PRs.

But this PR does define new ishim APIs for context aware buffer
allocation so shim implementation can be changed now to essentially
ignore the slotidx encoded in the grpid.

Ultimately all XRT applications should construct a hardware context
even in on legcacy platforms where only one physical slot is
supported.  The hwctx should be used explicitly in the kernel and bo
constructors.  Shim level APIs operating on hwctx or using hwctx will
be passed a hwctx_handle.  The hwctx_handle is currently typedef as a
uint8_t for the urpose of encoding in flags, but once encoding is no
longer needed the hwctx_handle can be changed to some opaque handle
type.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>